requirements.txt
edc-quota
edc-consent
edc-base
edc-constants
edc-audit
edc-device
edc-admin-exclude
django-revision
pytz
Django==1.6.7
Fabric==1.9.0
GitPython==0.3.2.RC1
M2Crypto==0.22.3
MySQL-python==1.2.5
South==1.0
Unipath==1.0
Werkzeug==0.9.6
argparse==1.2.1
async==0.6.1
django-dajax==0.9.2
django-dajaxice==0.6
django-databrowse==1.3
django-extensions==1.3.8
django-storages==1.1.8
django-tastypie==0.11.1
factory-boy==2.4.1
geo-utils==0.1
geopy==0.99
gitdb==0.5.4
gitflow==0.5.1
ipython==1.2.1
logger==1.4
nose==1.3.3
numpy==1.8.1
paramiko==1.14.0
pycups==1.9.67
pyodbc==3.0.7
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.3
smmap==0.8.2
wsgiref==0.1.2